[API update] String() - new parameter @samples
Add the new parameter @samples to allow the parameter @values get a rigid semantics.
If @values is provided, it is considered to be an exhaustive list of possible values.
If @samples is provided instead, then fuddly considers a value not in @samples to be valid
if it is compliant with the other parameters like @alphabet and size-related parameters.