API change: Operator() is now renamed Director()
Perform a renaming of Operator class to Director. The implementation is now
stored within director_helpers.py (where other renaming has been done).
And the various methods referring to Operators from plumbing.py, scenario.py
are now rebranded with "director" within. Documentation is also updated.

This renaming is because Disruptors is about to be renamed Operators.