Load sample files from current project to populate associated DMs
When a project is loaded, the data models associated with the project will be loaded
with the sample files present in the project.
If other samples are existing whether in the data model module itself or
in "<fuddly_data_folder>/imported_data", they will also be leveraged but with a
lower priority (meaning that if a file has the same filename as a file within the
project samples, the latter will be used).