[DM] New keyword to customize import from external data model
In a DM, when using 'import_from', it is now possible to specify a boolean value to
the new keyword 'samples_from_current_dm'.

It is useful in the case the imported data model leverages sample files (like the JSON DM)
to instruct whether to use the samples from the imported data model or to use instead samples
from the current data model. (Set to ``False`` by default.)